src/functions/level/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
black==22.3.0
CacheControl==0.12.11
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
cloudevents==1.2.0
deprecation==2.1.0
discord==1.7.3
discord.py==1.7.3
firebase-admin==4.5.1
Flask==2.2.0
functions-framework==3.0.0
google-api-core==2.8.2
google-api-python-client==2.47.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.0
google-cloud-firestore==2.4.0
google-cloud-pubsub==2.13.4
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.46.1
grpcio-status==1.46.1
gunicorn==20.1.0
html2image==2.0.1
htmlwebshot==0.1.2
httplib2==0.20.4
idna==3.3
install==1.3.5
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.3
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
proto-plus==1.20.3
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
rsa==4.8
six==1.16.0
tomli==2.0.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
watchdog==1.0.2
Werkzeug==2.2.2
yarl==1.7.2
